refactor(web): move shoelace config fo use in individual views
