+++ web/src/apps/main.css
@@ -4,15 +4,3 @@ body {
   text-rendering: optimizeLegibility;
   -webkit-font-smoothing: antialiased;
 }
-
-/*
- * Customize Shoelace
-  *
- * See https://shoelace.style/getting-started/customizing
- */
-:root {
-  /* Shoelace's default prefixed with Lato */
-  --sl-font-sans: Inter, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
-    Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji',
-    'Segoe UI Symbol';
-}
