+++ web/src/shoelace.css
@@ -1,0 +1,11 @@
+/*
+ * Customize Shoelace
+ *
+ * See https://shoelace.style/getting-started/customizing
+ */
+:root {
+  /* Shoelace's default prefixed with Inter */
+  --sl-font-sans: Inter, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
+    Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji',
+    'Segoe UI Symbol';
+}
